Change index.js to use modules.
